% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/System Regressor.R
\name{GeneratePythonScript}
\alias{GeneratePythonScript}
\title{Generate a TensorFlow Python script}
\usage{
GeneratePythonScript(epochs = 1e+05, learning.rate = 1e-04)
}
\arguments{
\item{epochs}{number of epochs to use in the model}

\item{learning.rate}{learning rate to use in the model}
}
\description{
Generate a TensorFlow Python script
}

