% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemRegressor.R
\name{GetSerializedTFData}
\alias{GetSerializedTFData}
\title{Get serialized data values back following python execution, internal function used by deepLearnR}
\usage{
GetSerializedTFData(file.name)
}
\arguments{
\item{file.name}{file name from which to read data}
}
\description{
Get serialized data values back following python execution, internal function used by deepLearnR
}

