% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/System Regressor.R
\name{SerializeData}
\alias{SerializeData}
\title{Serialize data to JSON and write to file}
\usage{
SerializeData(data, file.name)
}
\arguments{
\item{data}{data to serialize}

\item{file.name}{output file name}
}
\description{
Serialize data to JSON and write to file
}

