% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemRegressor.R
\name{SerializeTFData}
\alias{SerializeTFData}
\title{Serialize data to JSON and write to file, internal function used by deepLearnR}
\usage{
SerializeTFData(data, file.name)
}
\arguments{
\item{data}{data to serialize}

\item{file.name}{output file name}
}
\description{
Serialize data to JSON and write to file, internal function used by deepLearnR
}

