% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TensorFlowClassifier.R
\name{TensorFlow.predict}
\alias{TensorFlow.predict}
\title{Predict using a model(modelTag) the Yvalues for the X Matrix}
\usage{
TensorFlow.predict(modelTag, XTest, YTest)
}
\arguments{
\item{modelTag}{Tag for this model - referenced in the model ceate calls like TensorFlow.Classifier}

\item{XTest}{The X Matrix for test or prediction}

\item{YTest}{The Y Matrix for test, to calculate the accuracy}
}
\description{
Predict using a model(modelTag) the Yvalues for the X Matrix
}
\seealso{
\code{\link{TensorFlow.Classifier}}
}

