% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNNRegressor.R
\name{TensorFlow.regressorEval}
\alias{TensorFlow.regressorEval}
\title{Predict using a model(modelTag) the Yvalues for the X Matrix}
\usage{
TensorFlow.regressorEval(modelTag, calculateMSE = TRUE, calculateR2 = TRUE)
}
\arguments{
\item{modelTag}{Tag for this model - referenced in the model ceate calls like TensorFlow.Classifier}

\item{calculateMSE}{Yes/No to calculate the MSE}

\item{calculateR2}{Yes/No to compute R2}
}
\description{
Predict using a model(modelTag) the Yvalues for the X Matrix
}

