% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemRegressor.R
\name{TensorFlow.SystemLinReg}
\alias{TensorFlow.SystemLinReg}
\title{TensoFlow linear regression implementation}
\usage{
TensorFlow.SystemLinReg(X, Y, epochs = 1e+05, learning.rate = 1e-04)
}
\arguments{
\item{X}{the dependent variables in the model (currently only supports 1-D numeric vector)}

\item{Y}{the independent variable in the model}

\item{epochs}{number of epochs to use in the model}

\item{learning.rate}{learning rate to use in the model}
}
\description{
TensoFlow linear regression implementation
}
\examples{
{
x.vals <- seq(1:100)
y.vals <- 0.3 + 0.5 * x.vals
lm.tf.fit <- deepLearnR::TensorFlow.SystemLinReg(X = x.vals, Y = y.vals, 
                          epochs = 100000, learning.rate = .00005)
lm.tf.fit
}

}

