% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNNRegressor.R
\name{TensorFlowDNNRegressor}
\alias{TensorFlowDNNRegressor}
\title{Create  Classifier model based on the parameters}
\usage{
TensorFlowDNNRegressor(modelTag, X, y, test_size = 0.2, steps = 5000,
  learningRate = 0.1, batchSize = 1)
}
\arguments{
\item{modelTag}{Tag for this model - can be referenced in other calls like prediction}

\item{X}{The X Matrix for training}

\item{y}{The y Matrix for training}

\item{test_size}{The division of the dataset into training vs test}

\item{steps}{Number of epochs for training}

\item{learningRate}{The learning rate for optimize algorithm}

\item{batchSize}{Batch Size for the mini batch for optimization algorithms like SGD}
}
\description{
Create  Classifier model based on the parameters
}
\examples{
{
library(MASS)
data(Boston)
X <- Boston[,2:14]
y <- Boston[,1]

TensorFlowDNNRegressor(modelTag="tfdnnr-01", X=X, y=y, steps=5000)
pred <- TensorFlow.regressorEval(modelTag="tfdnnr-01")
mse <- rPython::python.get("mse")
r2 <- rPython::python.get("r2")
}

}

