% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanic.data.R
\docType{data}
\name{titanic.data}
\alias{titanic.data}
\title{The titanic Dataset}
\description{
A test data set of 12 variables 
\itemize{
  \item PassengerId. Id of the passenger  
  \item Survived. Whether they survived or not 
  \item Pclass. Class (1,2,3)
  \item Name. Name of the passenger
  \item Sex.
  \item Age.
  \item Sibsp. Number of Siblings/Spouses Aboard
  \item Parch. Number of Parents/Children Aboard
  \item Ticket. Ticket Number
  \item Fare. 
  \item Cabin. 
  \item Embarked. Port of Embarkation
}
}
\keyword{data}

