% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/System Regressor.R
\name{TensorFlow.SystemLinReg}
\alias{TensorFlow.SystemLinReg}
\title{Run a TensorFlow linear regression}
\usage{
TensorFlow.SystemLinReg(X, Y, epochs = 1e+05, learning.rate = 1e-04)
}
\arguments{
\item{X}{the dependent variables in the model}

\item{Y}{the independent variable in the model}

\item{epochs}{number of epochs to use in the model}

\item{learning.rate}{learning rate to use in the model}
}
\description{
Run a TensorFlow linear regression
}
\examples{
{
set.seed(123)
x.vals <- seq(1:100)
y.vals <- 0.3 + 0.5*(rnorm(n = 100, mean = 10, sd = 5) + x.vals)
lm.tf.fit <- TensorFlow.SystemLinReg(X = x.vals, Y = y.vals, epochs = 100000, learning.rate = .5)
lm.tf.fit
}

}

